fixed GCC warnings [Michael D. Adams]
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1661 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_dialogs.cpp
@@ -1276,11 +1276,11 @@ void get_net_color(
 	switch(index)
 	{
 		case _suicide_color:
-			color->red= color->green= LONG_MAX;
+			color->red= color->green= SHRT_MAX;
 			color->blue= 0;
 			break;
 		case _kill_color:
-			color->red= LONG_MAX;
+			color->red= SHRT_MAX;
 			color->green= color->blue= 0;
 			break;
 		case _death_color:
